fix: lint, imports and leverage common
