+++ jams-core/src/model_store/mod.rs
@@ -1,5 +1,5 @@
+pub mod azure_blob_storage;
+pub mod common;
 pub mod local;
 pub mod s3;
 pub mod storage;
-pub mod azure_blob_storage;
-pub mod common;
